base64, cat: fix build for Rust alpha

+++ src/cksum/cksum.rs
@@ -28,11 +28,11 @@ static VERSION: &'static str = "1.0.0";
 
 #[inline]
 fn crc_update(crc: u32, input: u8) -> u32 {
-    (crc << 8) ^ CRC_TABLE[((crc >> 24) as uint ^ input as uint) & 0xFF]
+    (crc << 8) ^ CRC_TABLE[((crc >> 24) as usize ^ input as usize) & 0xFF]
 }
 
 #[inline]
-fn crc_final(mut crc: u32, mut length: uint) -> u32 {
+fn crc_final(mut crc: u32, mut length: usize) -> u32 {
     while length != 0 {
         crc = crc_update(crc, length as u8);
         length >>= 8;
@@ -42,7 +42,7 @@ fn crc_final(mut crc: u32, mut length: uint) -> u32 {
 }
 
 #[inline]
-fn cksum(fname: &str) -> IoResult<(u32, uint)> {
+fn cksum(fname: &str) -> IoResult<(u32, usize)> {
     let mut crc = 0u32;
     let mut size = 0u;
 
@@ -74,7 +74,7 @@ fn cksum(fname: &str) -> IoResult<(u32, uint)> {
     }
 }
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
         getopts::optflag("V", "version", "output version information and exit"),
